[DO NOT REVIEW] improve: mv parts of daemon/containernetwork to daemon/utils and constants
