+++ .gitignore
@@ -19,7 +19,7 @@ bin
 # Dependency directories (remove the comment below to include it)
 vendor/
 
-config/
+/config/
 
 # only for kubebuilder
 main.go
