+++ pkg/constants/nic.go
@@ -1,0 +1,9 @@
+package constants
+
+const (
+	ContainerHostLinkPrefix = "hybr"
+	ContainerHostLinkMac    = "ee:ee:ee:ee:ee:ee"
+	ContainerInitLinkSuffix = "_c"
+	VxlanLinkInfix          = ".vxlan"
+	ContainerNicName        = "eth0"
+)
