+++ pkg/daemon/server/handle.go
@@ -23,6 +23,8 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/alibaba/hybridnet/pkg/daemon/utils"
+
 	"github.com/go-logr/logr"
 
 	corev1 "k8s.io/api/core/v1"
@@ -34,7 +36,6 @@ import (
 	networkingv1 "github.com/alibaba/hybridnet/pkg/apis/networking/v1"
 	"github.com/alibaba/hybridnet/pkg/constants"
 	daemonconfig "github.com/alibaba/hybridnet/pkg/daemon/config"
-	"github.com/alibaba/hybridnet/pkg/daemon/containernetwork"
 	"github.com/alibaba/hybridnet/pkg/daemon/controller"
 	"github.com/alibaba/hybridnet/pkg/request"
 
@@ -79,7 +80,7 @@ func (cdh *cniDaemonHandler) handleAdd(req *restful.Request, resp *restful.Respo
 	var netID *int32
 	var affectedIPInstances []*networkingv1.IPInstance
 
-	allocatedIPs := map[networkingv1.IPVersion]*containernetwork.IPInfo{
+	allocatedIPs := map[networkingv1.IPVersion]*utils.IPInfo{
 		networkingv1.IPv4: nil,
 		networkingv1.IPv6: nil,
 	}
@@ -160,7 +161,7 @@ func (cdh *cniDaemonHandler) handleAdd(req *restful.Request, resp *restful.Respo
 					return
 				}
 
-				allocatedIPs[networkingv1.IPv4] = &containernetwork.IPInfo{
+				allocatedIPs[networkingv1.IPv4] = &utils.IPInfo{
 					Addr: containerIP,
 					Gw:   gatewayIP,
 					Cidr: cidrNet,
@@ -172,7 +173,7 @@ func (cdh *cniDaemonHandler) handleAdd(req *restful.Request, resp *restful.Respo
 					return
 				}
 
-				allocatedIPs[networkingv1.IPv6] = &containernetwork.IPInfo{
+				allocatedIPs[networkingv1.IPv6] = &utils.IPInfo{
 					Addr: containerIP,
 					Gw:   gatewayIP,
 					Cidr: cidrNet,
@@ -291,7 +292,7 @@ func (cdh *cniDaemonHandler) errorWrapper(err error, status int, resp *restful.R
 	})
 }
 
-func printAllocatedIPs(allocatedIPs map[networkingv1.IPVersion]*containernetwork.IPInfo) string {
+func printAllocatedIPs(allocatedIPs map[networkingv1.IPVersion]*utils.IPInfo) string {
 	ipAddresseString := ""
 	if allocatedIPs[networkingv1.IPv4] != nil && allocatedIPs[networkingv1.IPv4].Addr != nil {
 		ipAddresseString = ipAddresseString + allocatedIPs[networkingv1.IPv4].Addr.String()
